User auth und Access Controll hinzugefügt

+++ server/services/attendance.service.js
@@ -1,23 +1,49 @@
 const httpStatus = require('http-status');
-const {Attendance} = require('../models');
+const { Attendance } = require('../models');
 const ApiError = require('../utils/ApiError');
+const mongoose = require('mongoose')
 
 /**
  * Get all attendance lists.
  * @returns {Promise<[Attendance]>}
  */
-const getAttendance = async () => {
-    return Attendance.find({});
+const getAttendance = async (user) => {
+    //user ist der Benutzer, der die Daten requestet
+
+    if (user.role === 'admin') {
+        //admin hat Zugriff auf alle Attendance Lists
+        return Attendance.find({})
+    } else if (user.role === 'trainer') {
+        //Wenn user ein Trainer o.ä. ist, wird für alle Attendance Lists die einer Gruppe zugewiesen sind, 
+        //auf die der User Zugriff hat
+        const list = await Attendance.find({'group._id': {$in: user.accessible_groups}})
+        
+        if(!list.length){
+            throw new ApiError(httpStatus.NOT_FOUND, 'No attendance list found for groups to which the user has access')
+        } else {
+            return list
+        }
+    } else {
+        throw new ApiError(httpStatus.UNAUTHORIZED, "The user's role has no access to any attendance list")
+    }
 };
 
-/**
+//WARNING Unused
+/*
  * Get a attendance list by ID.
  * @param {ObjectId} id
  * @returns {Promise<Attendance>}
  */
-const getAttendanceById = async (id) => {
-    return Attendance.findById(id);
+const getAttendanceById = async (user, id) => {
+    if (user.role === 'admin') {
+        //Admin kann auf alle Attendance Lists zugreifen
+        return Attendance.findById(id)
+    } else if (user.role === 'trainer'){
+        const list = await Attendance.findById(id);
+        return (list.access.includes(new mongoose.Types.ObjectId(id))) ? list : `The user has no access to the attendance list ${id}`
+    }
 };
+
 
 /**
  * Get a attendance list by groupID & date.
@@ -25,16 +51,17 @@ const getAttendanceById = async (id) => {
  * @param {Date} date Date of the searched attendance list
  * @returns {Promise<Attendance>}
  */
-const getAttendanceByDate = async (groupID, date) => {
-    const attendance = await getAttendanceByGroup(groupID)
-    for (const training of attendance.trainingssession) {
-        const dbDate = new Date(training.date)
-        date = new Date(date)
-        if (dbDate.getMonth() === date.getMonth() && dbDate.getDate() === date.getDate() && dbDate.getFullYear() === date.getFullYear()) {
-            return training
-        }
+const getTrainingssession = async (user, groupID, date) => {
+    //INFO Access controle wird von 'getAttendanceByGroup' gehandelt
+    const attendance = await getAttendanceByGroup(user, groupID)
+
+    const session = attendance.trainingssessions.find(element => element.date.toJSON() === new Date(date).toJSON())
+
+    if (typeof session !== 'undefined'){
+        return session
+    }else{
+        throw new ApiError(httpStatus.NOT_FOUND, "Requested Trainingssession not found")
     }
-    return {error: "No entrance for given date"}
 };
 
 /**
@@ -42,14 +69,20 @@ const getAttendanceByDate = async (groupID, date) => {
  * @param {ObjectID} groupID ID of the searched group
  * @returns {Promise<Attendance>}
  */
-const getAttendanceByGroup = async (groupID) => {
-    const allAttendance = await getAttendance()
-    for (const attendance of allAttendance) {
-        if (attendance.group.id === groupID) {
-            return Attendance(attendance)
+const getAttendanceByGroup = async (user, groupID) => {
+    if(user.role === 'admin'){
+        //Admin kann auf alle Listen zugreifen
+        return Attendance.findOne({'group._id': groupID})
+    }else if (user.role === 'trainer'){
+        //User kann nur auf Liste zugreifen, wenn er Zugriff auf die verbundene Gruppe hat
+        if(user.accessible_groups.includes(groupID)){
+            return Attendance.findOne({'group._id': groupID})
+        }else{
+            throw new ApiError(httpStatus.FORBIDDEN, "The user has no access to the group associated with this attendance list")
         }
+    }else{
+        throw new ApiError(httpStatus.UNAUTHORIZED, "The user's role has no access to attendance lists")
     }
-    return {error: "No entrance for given groupID"}
 };
 
 /**
@@ -57,8 +90,12 @@ const getAttendanceByGroup = async (groupID) => {
  * @param {Object} attendanceBody
  * @returns {Promise<Attendance>}
  */
-const createAttendance = async (attendanceBody) => {
-    return Attendance.create(attendanceBody);
+const createAttendance = async (user, attendanceBody) => {
+    if (user.role === 'admin') {
+        return Attendance.create(attendanceBody)
+    } else {
+        throw new ApiError(httpStatus.FORBIDDEN, "The user is not permitted to create a new attendance list")
+    }
 };
 
 /**
@@ -67,16 +104,21 @@ const createAttendance = async (attendanceBody) => {
  * @param sessionBody The body muss include date: Date, participants: [Member]
  * @returns {Promise<Attendance>}
  */
-const addTrainingssession = async (groupID, sessionBody) => {
-
-    const groupObj = await getAttendanceByGroup(groupID)
-    let sessions = groupObj.trainingssession
-
-    sessions.push(sessionBody)
-
-    return Attendance.findByIdAndUpdate({'_id': groupObj.id}, {'$set': {'trainingssession': sessions}})
+const addTrainingssession = async (user, groupID, sessionBody) => {
+    if (user.role === 'admin') {
+        return Attendance.findOneAndUpdate({ 'group._id': groupID}, { $addToSet: { trainingssessions: sessionBody } })
+    } else if (user.role === 'trainer') {
+        if (user.accessible_groups.includes(groupID)){
+            return Attendance.findOneAndUpdate({ 'group._id': groupID}, { $addToSet: { trainingssessions: sessionBody } })
+        } else{
+            throw new ApiError(httpStatus.FORBIDDEN, "The user is not permitted to add a trainingssession")
+        }
+    } else {
+        throw new ApiError(httpStatus.UNAUTHORIZED, "The user's role has no access to attendance lists")
+    }
 };
 
+//TODO Über diesen API Endpoint dürfen nicht Namen etc geändert werden Code so anpassen, dass nur update vom boolean möglich ist
 /**
  * Update a trainings session
  * @param groupID
@@ -84,18 +126,30 @@ const addTrainingssession = async (groupID, sessionBody) => {
  * @param {Object} sessionBody The body muss include date: Date, participants: [Member]
  * @returns {Promise<Attendance>}
  */
-const updateTrainingssession = async (groupID, date, sessionBody) => {
-    const groupObj = await getAttendanceByGroup(groupID)
-    let sessions = groupObj.trainingssession
-    date = new Date(date)
+const updateTrainingssession = async (user, groupID, date, sessionBody) => {
+    const session = await getTrainingssession(user, groupID, date)
 
-    for (let i = 0; i < sessions.length; i++) {
-        if (sessions[i].date.getMonth() === date.getMonth() && sessions[i].date.getDate() === date.getDate() && sessions[i].date.getFullYear() === date.getFullYear()){
-            sessions[i] = sessionBody
+    sessionBody.participants.forEach(participant => {
+        const temp = session.participants.find(foo => foo._id == participant._id)
+        if(typeof temp === 'undefined'){
+            session.participants.push(participant)
+        }else{
+            temp.attended = participant.attended
         }
+    })
+   
+    if(user.role === 'admin'){
+        return  Attendance.findOneAndUpdate({'group._id': groupID, 'trainingssessions.date': date}, {'$set': {'trainingssessions.$': session}})
+    } else if(user.role === 'trainer'){
+        if(user.accessible_groups.includes(groupID)){
+            return  Attendance.findOneAndUpdate({'group._id': groupID, 'trainingssessions.date': date}, {'$set': {'trainingssessions.$': session}})
+        }else{
+            throw new ApiError(httpStatus.FORBIDDEN, "The user is not permitted to update this trainingssession")
+        }
+    } else {
+        throw new ApiError(httpStatus.UNAUTHORIZED, "The user's role has no access to attendance lists")
     }
-
-    return Attendance.findByIdAndUpdate({'_id': groupObj.id}, {'$set': {'trainingssession': sessions}})
+    
 };
 
 /**
@@ -103,8 +157,12 @@ const updateTrainingssession = async (groupID, date, sessionBody) => {
  * @param attendanceID
  * @returns {Promise<Attendance>}
  */
-const deleteAttendance = async (attendanceID) => {
-    return Attendance.findByIdAndDelete(attendanceID)
+const deleteAttendance = async (user, attendanceID) => {
+    if (user.role === 'admin'){
+        return Attendance.findByIdAndDelete(attendanceID)
+    } else {
+        throw new ApiError(httpStatus.FORBIDDEN, "The user is not permitted to delete a attendance list")
+    }
 };
 
 /**
@@ -112,18 +170,18 @@ const deleteAttendance = async (attendanceID) => {
  * @param attendanceID
  * @returns {Promise<Attendance>}
  */
-const deleteTrainingssession = async (groupID, date) => {
-    const groupObj = await getAttendanceByGroup(groupID)
-    let sessions = groupObj.trainingssession
-    date = new Date(date)
-
-    for (let i = 0; i < sessions.length; i++) {
-        if (sessions[i].date.getMonth() === date.getMonth() && sessions[i].date.getDate() === date.getDate() && sessions[i].date.getFullYear() === date.getFullYear()){
-            sessions.splice(i, 1)
+const deleteTrainingssession = async (user, groupID, date) => {
+    if (user.role === 'admin'){
+        return  Attendance.findOneAndUpdate({'group._id': groupID, }, {'$pull': {'trainingssessions': {'date': date}}})
+    } else if(user.role === 'trainer'){
+        if(user.accessible_groups.includes(groupID)){
+            return  Attendance.findOneAndUpdate({'group._id': groupID, }, {'$pull': {'trainingssessions': {'date': date}}})
+        } else {
+            throw new ApiError(httpStatus.FORBIDDEN, "The user is not permitted to delete this trainingssession")
         }
+    } else {
+        throw new ApiError(httpStatus.UNAUTHORIZED, "The user's role has no access to attendance lists")
     }
-
-    return Attendance.findByIdAndUpdate({'_id': groupObj.id}, {'$set': {'trainingssession': sessions}})
 };
 
 //TODO Add new Functions here
@@ -133,8 +191,8 @@ module.exports = {
     createAttendance,
     updateTrainingssession,
     deleteAttendance,
-    getAttendanceByDate,
-    addTrainingssession,
+    getTrainingssession,
     getAttendanceByGroup,
-    deleteTrainingssession
+    deleteTrainingssession,
+    addTrainingssession
 };
